added login error message; file structure mods
